Move GetCpuTime function to base.
