+++ src/s2/base/sysinfo.h
@@ -1,0 +1,42 @@
+// Copyright 2018 Google Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS-IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+#ifndef S2_BASE_SYSINFO_H_
+#define S2_BASE_SYSINFO_H_
+
+#if defined(_WIN32) || defined(_WIN64)
+#include <Windows.h>
+#else
+#include <sys/resource.h>   // for rusage, RUSAGE_SELF
+#include <sys/time.h>
+#endif
+
+double GetCpuTime() {
+#if defined(_WIN32) || defined(_WIN64)
+  FILETIME creation, exit, kernel, user;
+  int res = GetProcessTimes(GetCurrentProcess(), &creation,
+                            &exit, &kernel, &user);
+  S2_CHECK_EQ(res, 0);
+  // high and low combine to 64-bit int, time uses 100-nanosecond intervals
+  return static_cast<double>((static_cast<uint64_t>(user.dwHighDateTime) << 32)
+                            + static_cast<uint64_t>(user.dwLowDateTime)) / 1e7;
+#else
+  struct rusage ru;
+  S2_CHECK_EQ(getrusage(RUSAGE_SELF, &ru), 0);
+  return ru.ru_utime.tv_sec + ru.ru_utime.tv_usec / 1e6;
+#endif
+}
+
+#endif  // S2_BASE_SYSINFO_H_
